refacto - use a trait for repeated tool methods
