+++ tests/functional/CommonDBRelationTest.php
@@ -38,17 +38,19 @@ use CommonDBRelation;
 use Glpi\Exception\Http\AccessDeniedHttpException;
 use Glpi\Exception\ItemLinkException;
 use Glpi\Exception\RedirectException;
-use Glpi\Security\ReAuth\ReAuthManager;
 use Glpi\Tests\DbTestCase;
+use Glpi\Tests\Glpi\Security\ReAuth\ReAuthTestTrait;
 use Group;
 use Group_User;
 use User;
 
 class CommonDBRelationTest extends DbTestCase
 {
+    use ReAuthTestTrait;
+
     public function tearDown(): void
     {
-        unset($GLOBALS['GLPI_IS_COMMAND_LINE']);
+        $this->restoreWebContext();
         parent::tearDown();
     }
 
@@ -487,25 +489,4 @@ class CommonDBRelationTest extends DbTestCase
         ]);
     }
 
-    private function fakeWebContext(): void
-    {
-        $GLOBALS['GLPI_IS_COMMAND_LINE'] = false;
-        $_SERVER['REQUEST_SCHEME'] = 'https';
-        $_SERVER['HTTP_HOST']      = 'glpi.example.org';
-        $_SERVER['REQUEST_URI']    = '/front/group_user.form.php';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-        $_SERVER['HTTP_REFERER']   = 'https://glpi.example.org/front/group.form.php';
-        $_GET  = [];
-        $_POST = [];
-    }
-
-    private function setReauthenticated(bool $reauthenticated): void
-    {
-        $_SESSION['glpi_currenttime'] = date('Y-m-d H:i:s');
-        if ($reauthenticated) {
-            (new ReAuthManager())->authenticate();
-        } else {
-            unset($_SESSION['glpi_reauth_until']);
-        }
-    }
 }
